requirements.txt
httpx>=0.27
python-dotenv>=1.0
rich>=13.0
requests>=2.31
shodan>=1.31
semgrep>=1.0
detect-secrets>=1.4
pip-audit>=2.7
pytest>=8.0
